requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
gunicorn==20.1.0
Pillow==7.2.0
psycopg2-binary==2.9.1
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
